EventManager built which has a nice client interface for raising/handling events.
Started work on a Sound implementation with OpenAL but that's still a work in progress.

A lot of boost files are checked in here, it's only a few MB so I figured it wouldn't hurt. Makes it easier to pull from github and compile without needing to go chasing down library dependencies.